Move InternalAggregationTestCase to test-framework module in order to make use of it from other modules.
This is a partial backport of #24559
