+++ core/src/test/java/org/elasticsearch/search/aggregations/metrics/min/InternalMinTests.java
@@ -20,10 +20,9 @@
 package org.elasticsearch.search.aggregations.metrics.min;
 
 import org.elasticsearch.search.DocValueFormat;
-import org.elasticsearch.search.aggregations.Aggregations;
-import org.elasticsearch.search.aggregations.InternalAggregationTestCase;
 import org.elasticsearch.search.aggregations.ParsedAggregation;
 import org.elasticsearch.search.aggregations.pipeline.PipelineAggregator;
+import org.elasticsearch.test.InternalAggregationTestCase;
 
 import java.util.List;
 import java.util.Map;
